Fix pointer type cast errors from Visual Studio (#617)
Visual Studio reports more pointer type cast errors with 64-bit build.
This will cover the issue on the existing targets.

+++ BootloaderCommonPkg/Library/AhciLib/AhciBlkIo.c
@@ -160,7 +160,7 @@ AhciReadBlocks (
     return EFI_UNSUPPORTED;
   }
 
-  AtaDeviceData = AhciFindDeviceData (mAhciPrivateData, DeviceIndex, 0xFFFFFFFF);
+  AtaDeviceData = AhciFindDeviceData (mAhciPrivateData, (UINT32)DeviceIndex, 0xFFFFFFFF);
 
   if (AtaDeviceData == NULL) {
     return EFI_NOT_FOUND;
@@ -203,7 +203,7 @@ AhciWriteBlocks (
     return EFI_UNSUPPORTED;
   }
 
-  AtaDeviceData = AhciFindDeviceData (mAhciPrivateData, DeviceIndex, 0xFFFFFFFF);
+  AtaDeviceData = AhciFindDeviceData (mAhciPrivateData, (UINT32)DeviceIndex, 0xFFFFFFFF);
 
   if (AtaDeviceData == NULL) {
     return EFI_NOT_FOUND;
@@ -244,7 +244,7 @@ AhciGetMediaInfo (
     return EFI_UNSUPPORTED;
   }
 
-  AtaDeviceData = AhciFindDeviceData (mAhciPrivateData, DeviceIndex, 0xFFFFFFFF);
+  AtaDeviceData = AhciFindDeviceData (mAhciPrivateData, (UINT32)DeviceIndex, 0xFFFFFFFF);
 
   if (AtaDeviceData == NULL) {
     return EFI_NOT_FOUND;
